Add sanitize_label helper and use in placement
